src/app/app.component.spec.ts
import {ComponentFixture, TestBed} from '@angular/core/testing';
import {AppComponent} from './app.component';

describe('AppComponent', () => {
  let component: AppComponent;
  let fixture: ComponentFixture<AppComponent>;
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [AppComponent],
    }).compileComponents();

    fixture = TestBed.createComponent(AppComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });


  it('should render title', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, app');
  });


  it('should not accept empty string', () => {
    component.form.setValue({
      userName: '',
      roomName: ''
    })

    expect(component.form.valid).toEqual(false)
  })

  it('should enter product name', () => {
    component.form.setValue({
      userName: 'username',
      roomName: 'room name'
    })

    expect(component.form.valid).toEqual(true)
  })
});

